+++ src/ipc/createDesktopShortcut.d.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
+ */
+
+import { App } from '../main/apps';
+
+export const registerHandlerFromMain: (
+    onCreateDesktopShortcut: (app: App) => void
+) => void;
+
+export const sendFromRenderer: (app: App) => void;
